Made calculate age return int32 for series.

+++ tests/library/test_age.py
@@ -16,19 +16,37 @@ def test_calculate_age():
     pd_as_of_dt = pd.Series([date(2020, 4, 6)], dtype="datetime64[ns]")
 
     # pd.Series, pd.Series
-    assert_series_equal(calculate_age(pd_birth_dt, pd_as_of_dt, "ALB"), pd.Series([29]))
-    assert_series_equal(calculate_age(pd_birth_dt, pd_as_of_dt, "ANB"), pd.Series([30]))
-    assert_series_equal(calculate_age(pd_birth_dt, pd_as_of_dt, "ACB"), pd.Series([30]))
+    assert_series_equal(
+        calculate_age(pd_birth_dt, pd_as_of_dt, "ALB"), pd.Series([29], dtype="int32")
+    )
+    assert_series_equal(
+        calculate_age(pd_birth_dt, pd_as_of_dt, "ANB"), pd.Series([30], dtype="int32")
+    )
+    assert_series_equal(
+        calculate_age(pd_birth_dt, pd_as_of_dt, "ACB"), pd.Series([30], dtype="int32")
+    )
 
     # pd.Series, date
-    assert_series_equal(calculate_age(pd_birth_dt, as_of_dt, "ALB"), pd.Series([29]))
-    assert_series_equal(calculate_age(pd_birth_dt, as_of_dt, "ANB"), pd.Series([30]))
-    assert_series_equal(calculate_age(pd_birth_dt, as_of_dt, "ACB"), pd.Series([30]))
+    assert_series_equal(
+        calculate_age(pd_birth_dt, as_of_dt, "ALB"), pd.Series([29], dtype="int32")
+    )
+    assert_series_equal(
+        calculate_age(pd_birth_dt, as_of_dt, "ANB"), pd.Series([30], dtype="int32")
+    )
+    assert_series_equal(
+        calculate_age(pd_birth_dt, as_of_dt, "ACB"), pd.Series([30], dtype="int32")
+    )
 
     # pd.Series, pd.Timestamp
-    assert_series_equal(calculate_age(pd_birth_dt, ts_as_of_dt, "ALB"), pd.Series([29]))
-    assert_series_equal(calculate_age(pd_birth_dt, ts_as_of_dt, "ANB"), pd.Series([30]))
-    assert_series_equal(calculate_age(pd_birth_dt, ts_as_of_dt, "ACB"), pd.Series([30]))
+    assert_series_equal(
+        calculate_age(pd_birth_dt, ts_as_of_dt, "ALB"), pd.Series([29], dtype="int32")
+    )
+    assert_series_equal(
+        calculate_age(pd_birth_dt, ts_as_of_dt, "ANB"), pd.Series([30], dtype="int32")
+    )
+    assert_series_equal(
+        calculate_age(pd_birth_dt, ts_as_of_dt, "ACB"), pd.Series([30], dtype="int32")
+    )
 
     # date, date
     assert calculate_age(birth_dt, as_of_dt, "ALB") == 29
@@ -36,9 +54,15 @@ def test_calculate_age():
     assert calculate_age(birth_dt, as_of_dt, "ACB") == 30
 
     # date, pd.Series
-    assert_series_equal(calculate_age(birth_dt, pd_as_of_dt, "ALB"), pd.Series([29]))
-    assert_series_equal(calculate_age(birth_dt, pd_as_of_dt, "ANB"), pd.Series([30]))
-    assert_series_equal(calculate_age(birth_dt, pd_as_of_dt, "ACB"), pd.Series([30]))
+    assert_series_equal(
+        calculate_age(birth_dt, pd_as_of_dt, "ALB"), pd.Series([29], dtype="int32")
+    )
+    assert_series_equal(
+        calculate_age(birth_dt, pd_as_of_dt, "ANB"), pd.Series([30], dtype="int32")
+    )
+    assert_series_equal(
+        calculate_age(birth_dt, pd_as_of_dt, "ACB"), pd.Series([30], dtype="int32")
+    )
 
     # date, pd.Timestamp
     assert calculate_age(birth_dt, ts_as_of_dt, "ALB") == 29
@@ -51,9 +75,15 @@ def test_calculate_age():
     assert calculate_age(ts_birth_dt, ts_as_of_dt, "ACB") == 30
 
     # pd.Timestamp, pd.Series
-    assert_series_equal(calculate_age(ts_birth_dt, pd_as_of_dt, "ALB"), pd.Series([29]))
-    assert_series_equal(calculate_age(ts_birth_dt, pd_as_of_dt, "ANB"), pd.Series([30]))
-    assert_series_equal(calculate_age(ts_birth_dt, pd_as_of_dt, "ACB"), pd.Series([30]))
+    assert_series_equal(
+        calculate_age(ts_birth_dt, pd_as_of_dt, "ALB"), pd.Series([29], dtype="int32")
+    )
+    assert_series_equal(
+        calculate_age(ts_birth_dt, pd_as_of_dt, "ANB"), pd.Series([30], dtype="int32")
+    )
+    assert_series_equal(
+        calculate_age(ts_birth_dt, pd_as_of_dt, "ACB"), pd.Series([30], dtype="int32")
+    )
 
     # pd.Timestamp, date
     assert calculate_age(ts_birth_dt, as_of_dt, "ALB") == 29
